+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed 
 - Fixed bug in enclosing functions
 - Parallelized PHR find initial Cut
+- caching subtree size
 
 ## [0.0.3] - 2021-08-12
 ### Added